requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
pydantic>=2.7.0
streamlit>=1.36.0
python-dotenv>=1.0.0
pytest>=8.0.0

# Optional, for the architecture diagram and graph rendering
pygraphviz; sys_platform != "win32"
graphviz>=0.20.0

# For the deck generation pipeline (not needed at runtime)
matplotlib>=3.8.0
